Add files via upload
The provided convolutional neural network implements a skin lesion classification model using the EfficientNetB7 architecture. It leverages the HAM10000 dataset, consisting of dermatoscopic images and metadata, for training and validation. The model is built by removing the top layers of EfficientNetB7, freezing the base layers, and adding custom layers for classification. Data generators are used to efficiently load and preprocess the images and metadata. The model is trained using categorical cross-entropy loss and Adam optimizer, with a focus on achieving high accuracy. After training, the model is evaluated on the validation set to assess its performance. The code is designed to run in a GPU-enabled environment, such as Google Colab, to accelerate the training process. The repository provides instructions for setting up the dataset and running the code, along with reporting the validation loss and accuracy achieved by the trained model.